Add simple shading
There's now a directional light direction passed in, which it uses to
lighten the color of surfaces facing the light and darken the color of
surfaces facing away from it. It also darkens the color of surfaces
facing oblique to the camera, to get a bit of rim shading and add depth.