Calculate normal in render instead of returning it
Previously, the SdfResult included the surface normal on a hit. This
required that the SDF be able to easily calculate what the normal should
be, which is not always simple. Instead, the render function just takes
some extra samples and uses the gradient in the range to calculate a
surface normal, which works regardless of the shape of the geometry.

Also push the minimum ray step size into a constant, which is also used
to determine the sampling offset for the surface normal calculation.