Add difference SDF, and make SdfResult a struct
The difference SDF was tricky, because there is no way to build a well
formed result SDF when taking the difference of two SDFs. It's just not
possible to correctly know the distance to the nearest surface in the
resulting shape when it's intersecting the negative shape. So the best
option is to make sure that if it's badly formed, it's always in a way
that will not impact the correctness of the raycasting algorithm. As
long as the resulting distance is never larger than the actual distance,
this is true. It may just go through more steps.

However, the raycasting algorithm was using the SDF to calculate the
surface normal, for shading purposes. If the SDF is malformed in any
direction, that impacts this calculation. So the SdfResult was changed
to include the surface normal, and each SDF was updated to provide this.
Now, the raycasting algorithm no longer needs to calculat it, so it can
handle and malformed SDF, as long as it still holds to the constraint of
the distance never being larger than the real distance.

However, the difference algorithm also needs the normal and color of the
negative shape while not inside it, in order to be able to shade the
carved out region with the normal and color of the negative shape. It's
only a hit on the combined shape when it's outside the negative shape.
So the SdfResult was changed to no longer be an enum with Hit and Miss
states, but to instead just be a struct with a range, normal and color,
and encode the Hit/Miss state only in the sign of the range.