Add truncated cone and plane SDFs
The truncated cone is formed by taking the difference between a cone and
a plane. But to do this, a plane SDF was needed. Which is very simple.

Also make Color implement Clone and Copy, and add an example for the
truncated cone.

The plane is too simple to warrant an example. The output would just be
a solid color anyway.