Add capsule and mirror SDFs
The capsule SDF allows the specification of a different radius for each
end, and is constructed from the union of two spheres and a truncated
cone.

The mirror SDF mirrors another SDF on one axis. This is needed by the
capsule SDF for the case in which the top radius is larger than the
bottom radius.

Also includes an example.