Convert SDFs to a trait
Previously an SDF was just a function which took a point and returned
an SdfResult. Now, there is an Sdf trait, so an Sdf is now a struct that
implements that trait. It was a check function on it, which serves the
same purpose as the old function, taking a point and returning an
SdfResult.

The purpose of this is performance. Some SDFs had to do some amount of
up-front calculation, and in the old system, that calculation had to be
repeated for every step of every ray. The new system allows an Sdf type
to do that work when it's created, so there is less redundant
calculation.

This also required fianlly adding the Transform SDF, which transforms
the input point and the output normal of the provided SDF, to move it to
a new location or scale it.

This also allowed for the removal of the mirror SDFs, as this can now be
handled with a negative scale in the Transform SDF.

Fianlly, updated every single Sdf, the render function, and all the
examples to work this way.