Make TruncatedCone support inversion
Previously, TruncatedCone would choke if the top radius was larger than
the bottom radius. Now in that case, it simply inverts the cone. This
also allowed the Capsule code to be simplified, as it no longer had to
make a special case for this. Essentially, the special case was moved
into TruncatedCone instead, which is a better place for it.