Add an animation system
The anim_render function takes all the normal arguments for render,
plus a start time, end time and frame time (the amount of time per
frame), and instead of an sdf, it takes a function which takes a time
and returns an sdf. The time values are not in any particular unit, and
can be interpretted any way that is convenient by the animation
function.