Move PNG saving in examples into common code
Create a new common directory in the examples directory, and add a new
save_png file. This removes the couple lines of boilerplate every file
needed to save a PNG at the end, and the cost of a couple lines of
boilerplate to include the common module. But the new boilerplate, while
slightly longer, is much simpler than the old.