Convert SDFs to be second-order functions
SDFs were originally simple functions, which worked well enough but
meant there were some redundant calculations going on, because a lot of
calculation could be done once up front instead of being redone with
every step of every ray. So, the SDFs were converted to be a trait, so
they could be implemented by a struct that could pre-calculate anything
needed at creation time. The problem is that this meant a complicated
SDF with a deep hierarchy needed to encode that entire heirarchy in a
very complicated generic type signature, which proved impractical.

So the solution now is that each SDF is a second-order function. Calling
the function does any initial calculations it needs, and then constructs
a closure which closes over the results of those calculations, and
returns that closure. The closure is what actually gets passed in to the
render function.

This also has the benefit of greatly simplifying the code for the SDFs,
and slightly reducing the verbosity of the code that constructions
compound SDFs.