Add skeleton system
This system allows you to specify a skeleton as a list of bones, where
each bone has a transform and a parent bone. The transform represents
it's base position. The bones are specified as a HashMap, indexed by the
bone name, which is the same name used to specify parenthood.

Then you can specify a pose, which is a transform for each bone, for how
the bone has moved from it's starting position.

Then you can specify a character, which associates SDFs with bones.

When all of these are passed into the skeleton function, it transforms
all the SDFs by the combined transform of their appropriate bone, and
then combines them all in a union.