Fix a subtle glitch in capsule
Capsule used a truncated cone for the middle portion. The bottom of the
truncated cone previously used special logic to figure out of the ray
was below the cone. Now, there's a new infinite_cone primitive, which
doesn't have a bottom, and the cone shape merely takes the difference
between that and a plane to cut off the bottom.

This avoids a subtle graphics glitch which resulting from surface
normals right on the edge being off. Now, every pixel of the truncated
cone is either 100% on the side or 100% on the bottom.