Fix a bug in capsule when the radii are equal
The capsule uses a truncated cone for the middle, and a truncated cone
where the top and bottom radii are the same is not a truncated cone. The
truncated cone in this library is literally just the difference between
a cone and a plane, and it's impossible to construct a cone such that it
can be cut off with both radii being equal.

So, added the cylinder shape (in the previous commit) and now updated
capsule to use a cylinder for the middle instead when the radii are
equal.