Make skeleton pose assert on invalid transforms
If the pose was given a NaN transform, it would just result in an
extremely slow render. The NaN would propogate through all the raycast
math, such that each ray would end up taking the shortest step length
every iteration. This was very hard to debug. Now, it asserts on any
NaN transforms when the pose is constructed if any of the transforms
contain a NaN, and tells you which one.