Add bottom half of humanoid skeleton
These are utility functions for constructing and posing a humanoid
skeleton. So far, it only supports the bottom half, from the hips down.

Also add an example that makes a walk cycle with the humanoid skeleton
and the animation system.

Also update the launch.json, to add this example to it, and remove all
configurations for debugging unit tests on examples. Examples don't get
unit tests.