Make plane primitive more flexible
Previously the plane was always located on the X/Z plane, and had to be
moved around with transforms. Now, the plane is specified with a normal
and offset, per the plane equation. It can also be specified with a
normal and one point, or with 3 points, but both of those will use that
data to calculate a normal and offset.

The cone, cube, cylinder and truncated cone were all updated to use the
new plane primitive.