Clean up the leg IK code
The leg IK code previously returned the leg rotation as a Quat, followed
by a thigh angle and knee angle, as 32, and then a foot rotation as a
Quat. The leg rotation and thigh angle needed to be immediately combined
into one rotation. Now, it just returns hip rotation, knee rotation and
foot rotation, all as Quats.