Fix transform modifier, needed to be inverted
The transform modifier should have been applying the inverse transform
to the input position, because it's moving the ray position, which will
effectively cause the position of the SDF being raycast against to move
in the opposite direction.

This required adjusting every place that used the transform modifier,
which is scattered throughout the whole codebase.