Add arm IK, and clean up humanoid code in general
Simplified the leg IK to match the new arm IK. Added the new
build_torso_only_pose, which poses just the hips, midriff, chest, neck
and head. This can then be fed into the skeleton, and the final position
of the hips and shoulders can be sampled as a starting point for the IK
solvers.

The example was updated to have the arms swing with the walk cycle.